requirements.txt
# Django Core
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Web Scraping
beautifulsoup4==4.12.3
scrapy==2.11.1
requests==2.31.0
lxml==5.1.0
html2text==2024.2.26

# Vector Database
qdrant-client==1.7.3

# Embeddings & ML
sentence-transformers==2.3.1
torch==2.2.0
numpy==1.26.3

# Google Gemini API
google-generativeai==0.3.2

# Async Tasks
celery==5.3.6
redis==5.0.1
django-celery-results==2.5.1
django-celery-beat==2.6.0

# Utilities
python-dotenv==1.0.0
pillow==10.2.0
psycopg2-binary==2.9.9

# Development
python-decouple==3.8